Done AutodetectPort change on main menu.

+++ src/config.cs
@@ -118,31 +118,46 @@ public class Config
 		}
 	}
 	
-	public static void UpdateEncoderConfiguration(EncoderConfiguration EconfNew)
+	public static void UpdateField(string field, string text)
 	{
 		//adapted from
 		//http://stackoverflow.com/a/2401873
+				
+		string tempfile = Path.GetTempFileName();
 
-		try {
-			string tempfile = Path.GetTempFileName();
-			using (var writer = new StreamWriter(tempfile))
-				using (var reader = new StreamReader(UtilAll.GetConfigFileName()))
+		LogB.Information("UpdateField, field: " + field + ", text: " + text);
+		
+		if(! File.Exists(UtilAll.GetConfigFileName())) {
+			try {
+				using (var writer = new StreamWriter(tempfile))
 				{
-					while (! reader.EndOfStream) {
-						string line = reader.ReadLine();
-						if (line != "" && line[0] != '#') 
-						{
-							string [] parts = line.Split(new char[] {'_'});
-							if(parts.Length == 2 && parts[0] == "EncoderConfiguration")
-								line = "EncoderConfiguration=" + EconfNew.ToStringOutput(EncoderConfiguration.Outputs.SQL);
-						}
-
-						writer.WriteLine(line);
-					}
+					writer.WriteLine(field + "=" + text);
 				}
-			File.Copy(tempfile, UtilAll.GetConfigFileName(), true);
-		} catch {
-			LogB.Warning("Cannot write EncoderConfiguration at Config.UpdateEncoderConfiguration");
+				File.Copy(tempfile, UtilAll.GetConfigFileName(), true);
+			} catch {
+				LogB.Warning("Cannot write at Config.UpdateField");
+			}
+		} else {
+			try {
+				using (var writer = new StreamWriter(tempfile))
+					using (var reader = new StreamReader(UtilAll.GetConfigFileName()))
+					{
+						while (! reader.EndOfStream) {
+							string line = reader.ReadLine();
+							if (line != "" && line[0] != '#') 
+							{
+								string [] parts = line.Split(new char[] {'='});
+								if(parts.Length == 2 && parts[0] == field)
+									line = field + "=" + text;
+							}
+
+							writer.WriteLine(line);
+						}
+					}
+				File.Copy(tempfile, UtilAll.GetConfigFileName(), true);
+			} catch {
+				LogB.Warning("Cannot write at Config.UpdateField");
+			}
 		}
 	}
